Move FFI part for string lists into its own module
